requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pillow==10.1.0
python-multipart==0.0.6
pytest==7.4.3
hypothesis==6.92.1
httpx==0.25.2
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
numpy==1.24.3
slowapi==0.1.9
python-dotenv==1.0.0
